Add script to generate maximalocal.mac.template when using BUILD_FOR_MOODLE
The Maxima shipped with qtype_stack requires that the maximalocal.mac.template
updated whenver any change to units etc have been made. Previously updating
the template was a manual process that required visiting the qtype_stack
health check script, and copy pasting a specific seciton of the configuration
it generates in to the maximalocal.mac.template file.

This patch automates this process by grabbing the relevant script from the
qtype_stack plugin, hacking it to run, then using it to generate the new
configuration.

The static maximalocal.mac.template file is still left in this repo for the
case when NOT building for Moodle. In this case Maxima is taken from StackQuestion
and no modification to the maximalocal.mac.template is required.